+++ Assets/Custom RP/Runtime/CameraRenderer.cs
@@ -2,66 +2,22 @@
 using UnityEngine.Experimental.Rendering.RenderGraphModule;
 using UnityEngine.Rendering;
 
-public partial class CameraRenderer
+public class CameraRenderer
 {
 	public const float renderScaleMin = 0.1f, renderScaleMax = 2f;
 
-	public static readonly int
-		bufferSizeId = Shader.PropertyToID("_CameraBufferSize"),
-		colorAttachmentId = Shader.PropertyToID("_CameraColorAttachment"),
-		depthAttachmentId = Shader.PropertyToID("_CameraDepthAttachment"),
-		colorTextureId = Shader.PropertyToID("_CameraColorTexture"),
-		depthTextureId = Shader.PropertyToID("_CameraDepthTexture"),
-		sourceTextureId = Shader.PropertyToID("_SourceTexture"),
-		srcBlendId = Shader.PropertyToID("_CameraSrcBlend"),
-		dstBlendId = Shader.PropertyToID("_CameraDstBlend");
-
 	static readonly CameraSettings defaultCameraSettings = new();
-
-	static readonly bool copyTextureSupported =
-		SystemInfo.copyTextureSupport > CopyTextureSupport.None;
-
-	static readonly Rect fullViewRect = new(0f, 0f, 1f, 1f);
-
-	CommandBuffer buffer;
-
-	ScriptableRenderContext context;
-
-	public Camera camera;
-
-	CullingResults cullingResults;
 
 	readonly Lighting lighting = new();
 
 	readonly PostFXStack postFXStack = new();
 
-	bool useHDR, useScaledRendering;
-
-	public bool useColorTexture, useDepthTexture, useIntermediateBuffer;
-
-	Vector2Int bufferSize;
-
 	readonly Material material;
 
-	readonly Texture2D missingTexture;
-
-	public CameraRenderer(Shader shader)
-	{
+	public CameraRenderer(Shader shader) =>
 		material = CoreUtils.CreateEngineMaterial(shader);
-		missingTexture = new Texture2D(1, 1)
-		{
-			hideFlags = HideFlags.HideAndDontSave,
-			name = "Missing"
-		};
-		missingTexture.SetPixel(0, 0, Color.white * 0.5f);
-		missingTexture.Apply(true, true);
-	}
 
-	public void Dispose()
-	{
-		CoreUtils.Destroy(material);
-		CoreUtils.Destroy(missingTexture);
-	}
+	public void Dispose() => CoreUtils.Destroy(material);
 
 	public void Render(
 		RenderGraph renderGraph,
@@ -71,9 +27,6 @@ public partial class CameraRenderer
 		ShadowSettings shadowSettings, PostFXSettings postFXSettings,
 		int colorLUTResolution)
 	{
-		this.context = context;
-		this.camera = camera;
-
 		ProfilingSampler cameraSampler;
 		CameraSettings cameraSettings;
 		if (camera.TryGetComponent(out CustomRenderPipelineCamera crpCamera))
@@ -87,6 +40,7 @@ public partial class CameraRenderer
 			cameraSettings = defaultCameraSettings;
 		}
 
+		bool useColorTexture, useDepthTexture;
 		if (camera.cameraType == CameraType.Reflection)
 		{
 			useColorTexture = bufferSettings.copyColorReflection;
@@ -104,14 +58,27 @@ public partial class CameraRenderer
 		}
 
 		float renderScale = cameraSettings.GetRenderScale(bufferSettings.renderScale);
-		useScaledRendering = renderScale < 0.99f || renderScale > 1.01f;
-		PrepareForSceneWindow();
-		if (!Cull(shadowSettings.maxDistance))
+		bool useScaledRendering = renderScale < 0.99f || renderScale > 1.01f;
+
+#if UNITY_EDITOR
+		if (camera.cameraType == CameraType.SceneView)
+		{
+			ScriptableRenderContext.EmitWorldGeometryForSceneView(camera);
+			useScaledRendering = false;
+		}
+#endif
+
+		if (!camera.TryGetCullingParameters(
+			out ScriptableCullingParameters scriptableCullingParameters))
 		{
 			return;
 		}
+		scriptableCullingParameters.shadowDistance =
+			Mathf.Min(shadowSettings.maxDistance, camera.farClipPlane);
+		CullingResults cullingResults = context.Cull(ref scriptableCullingParameters);
 
-		useHDR = bufferSettings.allowHDR && camera.allowHDR;
+		bool useHDR = bufferSettings.allowHDR && camera.allowHDR;
+		Vector2Int bufferSize = default;
 		if (useScaledRendering)
 		{
 			renderScale = Mathf.Clamp(renderScale, renderScaleMin, renderScaleMax);
@@ -130,7 +97,7 @@ public partial class CameraRenderer
 			colorLUTResolution, cameraSettings.finalBlendMode,
 			bufferSettings.bicubicRescaling, bufferSettings.fxaa);
 
-		useIntermediateBuffer = useScaledRendering ||
+		bool useIntermediateBuffer = useScaledRendering ||
 			useColorTexture || useDepthTexture || postFXStack.IsActive;
 
 		var renderGraphParameters = new RenderGraphParameters
@@ -141,195 +108,50 @@ public partial class CameraRenderer
 			rendererListCulling = true,
 			scriptableRenderContext = context
 		};
-		buffer = renderGraphParameters.commandBuffer;
 		using (renderGraph.RecordAndExecute(renderGraphParameters))
 		{
 			using var _ = new RenderGraphProfilingScope(renderGraph, cameraSampler);
+
 			LightingPass.Record(
 				renderGraph, lighting,
 				cullingResults, shadowSettings, useLightsPerObject,
 				cameraSettings.maskLights ? cameraSettings.renderingLayerMask : -1);
-			SetupPass.Record(renderGraph, this);
+
+			CameraRendererTextures textures = SetupPass.Record(
+				renderGraph, useIntermediateBuffer, useColorTexture, useDepthTexture,
+				useHDR, bufferSize, camera);
 
 			GeometryPass.Record(
 				renderGraph, camera, cullingResults,
-				useLightsPerObject, cameraSettings.renderingLayerMask, true);
+				useLightsPerObject, cameraSettings.renderingLayerMask, true, textures);
 
-			SkyboxPass.Record(renderGraph, camera);
+			SkyboxPass.Record(renderGraph, camera, textures);
 
-			if (useColorTexture || useDepthTexture)
-			{
-				CopyAttachmentsPass.Record(renderGraph, this);
-			}
+			var copier = new CameraRendererCopier(
+				material, camera, cameraSettings.finalBlendMode);
+			CopyAttachmentsPass.Record(
+				renderGraph, useColorTexture, useDepthTexture, copier, textures);
 
 			GeometryPass.Record(
 				renderGraph, camera, cullingResults,
-				useLightsPerObject, cameraSettings.renderingLayerMask, false);
+				useLightsPerObject, cameraSettings.renderingLayerMask, false, textures);
 
 			UnsupportedShadersPass.Record(renderGraph, camera, cullingResults);
 
 			if (postFXStack.IsActive)
 			{
-				PostFXPass.Record(renderGraph, postFXStack);
+				PostFXPass.Record(renderGraph, postFXStack, textures);
 			}
 			else if (useIntermediateBuffer)
 			{
-				FinalPass.Record(renderGraph, this, cameraSettings.finalBlendMode);
+				FinalPass.Record(renderGraph, copier, textures);
 			}
-			GizmosPass.Record(renderGraph, this);
+			GizmosPass.Record(renderGraph, useIntermediateBuffer, copier, textures);
 		}
 
-		Cleanup();
-		Submit();
-		CommandBufferPool.Release(renderGraphParameters.commandBuffer);
-	}
-
-	bool Cull(float maxShadowDistance)
-	{
-		if (camera.TryGetCullingParameters(out ScriptableCullingParameters p))
-		{
-			p.shadowDistance = Mathf.Min(maxShadowDistance, camera.farClipPlane);
-			cullingResults = context.Cull(ref p);
-			return true;
-		}
-		return false;
-	}
-
-	public void Setup()
-	{
-		context.SetupCameraProperties(camera);
-		CameraClearFlags flags = camera.clearFlags;
-
-		if (useIntermediateBuffer)
-		{
-			if (flags > CameraClearFlags.Color)
-			{
-				flags = CameraClearFlags.Color;
-			}
-			buffer.GetTemporaryRT(
-				colorAttachmentId, bufferSize.x, bufferSize.y,
-				0, FilterMode.Bilinear, useHDR ?
-					RenderTextureFormat.DefaultHDR : RenderTextureFormat.Default);
-			buffer.GetTemporaryRT(
-				depthAttachmentId, bufferSize.x, bufferSize.y,
-				32, FilterMode.Point, RenderTextureFormat.Depth);
-			buffer.SetRenderTarget(
-				colorAttachmentId,
-				RenderBufferLoadAction.DontCare, RenderBufferStoreAction.Store,
-				depthAttachmentId,
-				RenderBufferLoadAction.DontCare, RenderBufferStoreAction.Store);
-		}
-
-		buffer.ClearRenderTarget(
-			flags <= CameraClearFlags.Depth,
-			flags <= CameraClearFlags.Color,
-			flags == CameraClearFlags.Color ?
-				camera.backgroundColor.linear : Color.clear);
-		buffer.SetGlobalTexture(colorTextureId, missingTexture);
-		buffer.SetGlobalTexture(depthTextureId, missingTexture);
-		buffer.SetGlobalVector(bufferSizeId, new Vector4(
-			1f / bufferSize.x, 1f / bufferSize.y, bufferSize.x, bufferSize.y));
-		ExecuteBuffer();
-	}
-
-	void Cleanup()
-	{
 		lighting.Cleanup();
-		if (useIntermediateBuffer)
-		{
-			buffer.ReleaseTemporaryRT(colorAttachmentId);
-			buffer.ReleaseTemporaryRT(depthAttachmentId);
-			if (useColorTexture)
-			{
-				buffer.ReleaseTemporaryRT(colorTextureId);
-			}
-			if (useDepthTexture)
-			{
-				buffer.ReleaseTemporaryRT(depthTextureId);
-			}
-		}
-	}
-
-	void Submit()
-	{
-		ExecuteBuffer();
+		context.ExecuteCommandBuffer(renderGraphParameters.commandBuffer);
 		context.Submit();
-	}
-
-	public void ExecuteBuffer()
-	{
-		context.ExecuteCommandBuffer(buffer);
-		buffer.Clear();
-	}
-
-	public void CopyAttachments()
-	{
-		ExecuteBuffer();
-		if (useColorTexture)
-		{
-			buffer.GetTemporaryRT(
-				colorTextureId, bufferSize.x, bufferSize.y,
-				0, FilterMode.Bilinear, useHDR ?
-					RenderTextureFormat.DefaultHDR : RenderTextureFormat.Default);
-			if (copyTextureSupported)
-			{
-				buffer.CopyTexture(colorAttachmentId, colorTextureId);
-			}
-			else
-			{
-				Draw(colorAttachmentId, colorTextureId);
-			}
-		}
-		if (useDepthTexture)
-		{
-			buffer.GetTemporaryRT(
-				depthTextureId, bufferSize.x, bufferSize.y,
-				32, FilterMode.Point, RenderTextureFormat.Depth);
-			if (copyTextureSupported)
-			{
-				buffer.CopyTexture(depthAttachmentId, depthTextureId);
-			}
-			else
-			{
-				Draw(depthAttachmentId, depthTextureId, true);
-			}
-		}
-		if (!copyTextureSupported)
-		{
-			buffer.SetRenderTarget(
-				colorAttachmentId,
-				RenderBufferLoadAction.Load, RenderBufferStoreAction.Store,
-				depthAttachmentId,
-				RenderBufferLoadAction.Load, RenderBufferStoreAction.Store);
-		}
-		ExecuteBuffer();
-	}
-
-	public void Draw(
-		RenderTargetIdentifier from, RenderTargetIdentifier to, bool isDepth = false)
-	{
-		buffer.SetGlobalTexture(sourceTextureId, from);
-		buffer.SetRenderTarget(
-			to, RenderBufferLoadAction.DontCare, RenderBufferStoreAction.Store);
-		buffer.SetViewport(camera.pixelRect);
-		buffer.DrawProcedural(
-			Matrix4x4.identity, material, isDepth ? 1 : 0, MeshTopology.Triangles, 3);
-	}
-
-	public void DrawFinal(CameraSettings.FinalBlendMode finalBlendMode)
-	{
-		buffer.SetGlobalFloat(srcBlendId, (float)finalBlendMode.source);
-		buffer.SetGlobalFloat(dstBlendId, (float)finalBlendMode.destination);
-		buffer.SetGlobalTexture(sourceTextureId, colorAttachmentId);
-		buffer.SetRenderTarget(
-			BuiltinRenderTextureType.CameraTarget,
-			finalBlendMode.destination == BlendMode.Zero && camera.rect == fullViewRect ?
-				RenderBufferLoadAction.DontCare : RenderBufferLoadAction.Load,
-			RenderBufferStoreAction.Store);
-		buffer.SetViewport(camera.pixelRect);
-		buffer.DrawProcedural(
-			Matrix4x4.identity, material, 0, MeshTopology.Triangles, 3);
-		buffer.SetGlobalFloat(srcBlendId, 1f);
-		buffer.SetGlobalFloat(dstBlendId, 0f);
+		CommandBufferPool.Release(renderGraphParameters.commandBuffer);
 	}
 }