+++ Assets/Custom RP/Runtime/Passes/FinalPass.cs
@@ -5,25 +5,27 @@ public class FinalPass
 {
 	static readonly ProfilingSampler sampler = new("Final");
 
-	CameraRenderer renderer;
+	CameraRendererCopier copier;
 
-	CameraSettings.FinalBlendMode finalBlendMode;
+	TextureHandle colorAttachment;
 
 	void Render(RenderGraphContext context)
 	{
-		renderer.DrawFinal(finalBlendMode);
-		renderer.ExecuteBuffer();
+		CommandBuffer buffer = context.cmd;
+		copier.CopyToCameraTarget(buffer, colorAttachment);
+		context.renderContext.ExecuteCommandBuffer(buffer);
+		buffer.Clear();
 	}
 
 	public static void Record(
 		RenderGraph renderGraph,
-		CameraRenderer renderer,
-		CameraSettings.FinalBlendMode finalBlendMode)
+		CameraRendererCopier copier,
+		in CameraRendererTextures textures)
 	{
 		using RenderGraphBuilder builder =
 			renderGraph.AddRenderPass(sampler.name, out FinalPass pass, sampler);
-		pass.renderer = renderer;
-		pass.finalBlendMode = finalBlendMode;
+		pass.copier = copier;
+		pass.colorAttachment = builder.ReadTexture(textures.colorAttachment);
 		builder.SetRenderFunc<FinalPass>((pass, context) => pass.Render(context));
 	}
 }