+++ Assets/Custom RP/Runtime/PostFXStack.cs
@@ -118,7 +118,7 @@ public partial class PostFXStack
 		ApplySceneViewState();
 	}
 
-	public void Render(RenderGraphContext context, int sourceId)
+	public void Render(RenderGraphContext context, TextureHandle sourceId)
 	{
 		buffer = context.cmd;
 		if (DoBloom(sourceId))
@@ -134,7 +134,7 @@ public partial class PostFXStack
 		buffer.Clear();
 	}
 
-	bool DoBloom(int sourceId)
+	bool DoBloom(RenderTargetIdentifier sourceId)
 	{
 		BloomSettings bloom = settings.Bloom;
 		int width, height;
@@ -309,7 +309,7 @@ public partial class PostFXStack
 			fxaa.fixedThreshold, fxaa.relativeThreshold, fxaa.subpixelBlending));
 	}
 
-	void DoFinal(int sourceId)
+	void DoFinal(RenderTargetIdentifier sourceId)
 	{
 		ConfigureColorAdjustments();
 		ConfigureWhiteBalance();